Add a base serial class, generic enough that it can be used for AUX or USB
My Idea is to create a SerialPort base class with a few virtual functions, Read/Write/Baud/Parity/Stops/DMA/Inverter. Which each driver will subclass USB or AUX.

On top of this layer will be a serial_interface. Which is going to be next level of networking. e.g. (FRSky BT Module, RAW(passthrough), Wifi (wroom) module, HM10, etc...) Which will be responsible for configuring the device and sending data.

On top of this layer will be a serial_function. Which is going to be Mavlink, Debug, Trainer, Telemetry, LUA.

Which should allow you to send any data over any interface. I'm just not sure how strapped each target is on memory/flash if this will work overall.